+++ src/date/util/props.js
@@ -16,6 +16,10 @@ const ALIASES = {
 const SHORTCUT_REG = /^(-?\d+(?:\.\d+)?) (year|month|week|day|hour|minute|(?:milli)?second)s?$/;
 
 const FRACTIONAL_UNITS = {
+  'date': {
+    unit: 'hour',
+    mult: 24,
+  },
   'day': {
     unit: 'hour',
     mult: 24,
@@ -31,6 +35,9 @@ const FRACTIONAL_UNITS = {
   'second': {
     unit: 'millisecond',
     mult: 1000,
+  },
+  'millisecond': {
+    mult: 1,
   }
 };
 
@@ -38,7 +45,7 @@ export function normalizeProps(props) {
   const normalized = {};
   for (let [unit, val] of Object.entries(props)) {
     unit = getNormalizedUnit(unit);
-    normalized[unit] = val;
+    normalizeFractionalUnit(normalized, unit, val);
   }
   return normalized;
 }
@@ -46,26 +53,29 @@ export function normalizeProps(props) {
 export function getPropsFromString(str) {
   const match = str.match(SHORTCUT_REG);
   if (!match) {
-    throwInvalidFormat(str);
+    throw new TypeError(`Invalid input ${str}.`);
   }
-  const params = {};
-  let val = parseFloat(match[1]);
-  const unit = match[2];
-  const fraction = val % 1;
-  if (fraction !== 0) {
-    const fUnit = FRACTIONAL_UNITS[unit];
-    if (!fUnit) {
-      throwInvalidFormat(str);
-    }
-    params[fUnit.unit] = Math.round(fraction * fUnit.mult);
-    val = Math.trunc(val);
-  }
-  params[unit] = val;
-  return params;
+  const props = {};
+  normalizeFractionalUnit(props, match[2], parseFloat(match[1]));
+  return props;
 }
 
-function throwInvalidFormat(str) {
-  throw new TypeError(`Invalid input ${str}.`);
+function normalizeFractionalUnit(props, unit, val) {
+  if (unit in FRACTIONAL_UNITS) {
+    const fraction = val % 1;
+    if (fraction) {
+      const { unit: fUnit, mult } = FRACTIONAL_UNITS[unit];
+      if (fUnit) {
+        props[fUnit] = Math.round(fraction * mult);
+        // Truncate 5 in 5.5 hours.
+        val = Math.trunc(val);
+      } else {
+        // Round milliseconds.
+        val = Math.round(val);
+      }
+    }
+  }
+  props[unit] = val;
 }
 
 function getNormalizedUnit(unit) {